third_party/WebKit/LayoutTests/fast/canvas/canvas-lose-restore-googol-size-expected.txt

Tests to ensure correct behaviour of canvas loss and restoration when size is extremely large then, restored to a reasonable value.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS contextLostTest is false
PASS ctx.isContextLost() is false
PASS contextLostTest is true
PASS ctx.isContextLost() is true
PASS contextLostTest is true
PASS ctx.isContextLost() is true
PASS contextLostTest is true
PASS ctx.isContextLost() is true
PASS successfullyParsed is true

TEST COMPLETE
PASS Graphics context lost event dispatched.
PASS contextLostTest is true
PASS ctx.isContextLost() is true
PASS Context restored event dispatched after context lost.
PASS contextLostTest is false
PASS ctx.isContextLost() is false
